+++ Classes/Main/MainController.h
@@ -109,7 +109,7 @@
 /// @name Callback
 
 - (IBAction) rename:(id)sender;
-- (IBAction)tagsFromFilename:(id)sender;
+- (IBAction) tagsFromFilename:(id)sender;
 
 /**  
  * Shows the sheet for searching for tags
@@ -119,6 +119,7 @@
 
 - (IBAction) reopen:(id)sender;
 
+- (IBAction) renumberFiles:(id)sender;